# solution/docs/swagger.yaml
components:
  schemas:
...
    dto.ModerationBody:
      properties:
        moderation:
          type: boolean
      required:
      - moderation
      type: object
    dto.ModerationVerdictBody:
      properties:
        verdict:
          type: boolean
      required:
      - verdict
      type: object
    dto.NeuroBody:
      properties:
        ad_title:
          type: string
      required:
      - ad_title
      type: object
...
    handlers.NeuroAnswer:
      properties:
        ad_text:
          type: string
      type: object
    handlers.UploadedImageKey:
      properties:
        key:
          type: string
      type: object
...
    models.Campaign:
      properties:
        ad_text:
          type: string
        ad_title:
          type: string
        advertiser_id:
          type: string
        campaign_id:
          type: string
        clicks_limit:
          type: integer
        cost_per_click:
          type: number
        cost_per_impression:
          type: number
        end_date:
          type: integer
        image:
          type: string
        impressions_limit:
          type: integer
        moderated:
          type: boolean
        start_date:
          type: integer
        targeting:
          $ref: '#/components/schemas/models.Targeting'
      required:
      - ad_text
      - ad_title
      - clicks_limit
      - cost_per_click
      - cost_per_impression
      - end_date
      - impressions_limit
      - start_date
      type: object
...
    models.Client:
      properties:
        age:
          type: integer
        client_id:
          type: string
        gender:
          enum:
          - MALE
          - FEMALE
          type: string
        location:
          type: string
        login:
          type: string
      required:
      - age
      - client_id
      - gender
      - location
      - login
      type: object
...
    models.Targeting:
      properties:
        age_from:
          type: integer
        age_to:
          type: integer
        gender:
          enum:
          - MALE
          - FEMALE
          - ALL
          type: string
        location:
          type: string
      type: object
...
info:
  description: PROOOOOOOOOD
  title: Prod-2025 AD API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /advertisers/{id}/campaigns/{camp_id}/image:
    delete:
      description: Delete an image from a campaign
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Delete an image
      tags:
      - campaigns
    get:
      description: Get an image from a campaign
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            image/*:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get an image
      tags:
      - campaigns
    put:
      description: Upload an image to a campaign
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: Campaign ID
        in: path
        name: camp_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          image/*:
            schema:
              format: binary
              type: string
        description: Image file
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.UploadedImageKey'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Upload an image
      tags:
      - campaigns
...
  /clients/{id}:
    get:
      description: Get a client by client ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Client'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get a client
      tags:
      - clients
  /clients/bulk:
    post:
      description: Create multiple clients in bulk
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/models.Client'
              type: array
        description: List of clients
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Client'
                type: array
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
      summary: Bulk create clients
      tags:
      - clients
...
  /moderation:
    post:
      description: Enable or disable moderation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.ModerationBody'
        description: Moderation settings
        required: true
      responses:
        "200":
          content:
            application/json: {}
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Change moderation settings
      tags:
      - moderation
  /moderation/campaigns:
    get:
      description: Retrieve a list of unmoderated campaigns
      parameters:
      - description: Page size
        in: query
        name: size
        required: true
        schema:
          type: integer
      - description: Page number
        in: query
        name: page
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/models.Campaign'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
      summary: Get unmoderated campaigns
      tags:
      - moderation
  /moderation/campaigns/{id}:
    post:
      description: Approve or reject a campaign based on the verdict
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.ModerationVerdictBody'
        description: Moderation verdict
        required: true
      responses:
        "200":
          content:
            application/json: {}
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Moderate a campaign
      tags:
      - moderation
  /neuro/{id}:
    post:
      description: Generate an advertisement description using LLM service
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.NeuroBody'
        description: Neuro Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.NeuroAnswer'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Generate ad description
      tags:
      - neuro
...
servers:
- url: http://localhost:8080/